# openapi.yaml
...
openapi: 3.1.0
info:
  title: aperIA - Python API
  description: '

    API do **aperIA** — ASPM (Application Security Posture Management) open-source.


    Um webhook de Pull Request do GitHub dispara um pipeline de scanners de

    segurança em 3 tiers (via Celery). O Claude raciocina como atacante

    (correlaciona findings, monta attack paths) e entrega patches como

    **GitHub code suggestions** para aprovação humana.


    > **Premissa inviolável:** o aperIA **nunca** aplica código sozinho.

    > Todo patch sai como code suggestion para aprovação humana.

    '
  contact:
    name: aperIA
  license:
    name: Open Source
  version: 1.0.0
paths:
...
  /scans:
    get:
      tags:
      - scans
      summary: Listar scans recentes
      description: Lista os scans do usuário logado, paginados (mais recentes primeiro).
      operationId: list_scans_scans_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanJobPage'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /scans/{commit_sha}:
    get:
      tags:
      - scans
      summary: Consultar status de um scan por commit
      description: Retorna o status de um scan do usuário pelo `commit_sha`, com resumo de findings.
      operationId: get_scan_scans__commit_sha__get
      security:
      - HTTPBearer: []
      parameters:
      - name: commit_sha
        in: path
        required: true
        schema:
          type: string
          title: Commit Sha
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanJobResponse'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Scan nao encontrado.
          content:
            application/json:
              example:
                detail: Scan nao encontrado
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /scans/{commit_sha}/report:
    get:
      tags:
      - scans
      summary: Listar relatorios de um scan (todos os tiers)
      description: 'Retorna os relatorios (um por tier) do `commit_sha`, se for do usuário.


        404 se o scan não existir ou não pertencer ao usuário. Lista vazia se o

        scan existe mas ainda não há relatórios (pipeline em andamento).'
      operationId: get_scan_reports_scans__commit_sha__report_get
      security:
      - HTTPBearer: []
      parameters:
      - name: commit_sha
        in: path
        required: true
        schema:
          type: string
          title: Commit Sha
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanReportsResponse'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Scan nao encontrado.
          content:
            application/json:
              example:
                detail: Scan nao encontrado
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /scans/{commit_sha}/tiers/{tier}/report:
    get:
      tags:
      - scans
      summary: Consultar relatorio de um tier especifico
      description: Retorna o relatorio de um tier (1-3) do `commit_sha`, se for do usuário.
      operationId: get_scan_report_by_tier_scans__commit_sha__tiers__tier__report_get
      security:
      - HTTPBearer: []
      parameters:
      - name: commit_sha
        in: path
        required: true
        schema:
          type: string
          title: Commit Sha
      - name: tier
        in: path
        required: true
        schema:
          type: integer
          maximum: 3
          minimum: 1
          title: Tier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanReportResponse'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Relatorio nao encontrado.
          content:
            application/json:
              example:
                detail: Relatorio nao encontrado
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /repositories/{repository_id}/scan:
    post:
      tags:
      - repositories
      summary: Iniciar scan manual do repositório
      description: "Dispara, sob demanda, o **mesmo** pipeline que o webhook de PR dispara.\n\nO commit\
        \ analisado é o HEAD do `default_branch` do repositório, resolvido\nao vivo no GitHub com o installation\
        \ token. Como não há PR, o pipeline roda\ncom `pr_number=None`: os relatórios ficam na projeção\
        \ da API em vez de\nvirarem comentário de PR.\n\nContrato de erros:\n\n- `404` se o repositório\
        \ não existir **ou** não pertencer ao usuário (nunca\n  403 — não vazamos a existência do recurso).\n\
        - `409` se o repositório estiver `active=false`. Desativado significa \"não\n  analisar\": um\
        \ scan manual aqui seria contraditório, e responder 404\n  esconderia do usuário um repositório\
        \ que ele mesmo pode reativar via\n  `PATCH /repositories/{id}`. Por isso 409 com mensagem explícita.\n\
        - `409` se já existir um `ScanJob` em andamento (algum tier `queued`/\n  `running`) para o commit\
        \ resolvido — evita duplicar o pipeline.\n- `502` se o GitHub não responder o HEAD do branch.\n\
        - `503` se o GitHub App não estiver configurado (mesmo espírito do\n  `GET /github/connect`, mas\
        \ aqui o que falta são as credenciais do App —\n  `GITHUB_APP_ID` / `GITHUB_PRIVATE_KEY_PATH`\
        \ — necessárias para emitir o\n  installation token)."
      operationId: trigger_repository_scan_repositories__repository_id__scan_post
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Repository Id
      responses:
        '202':
          description: Pipeline disparado para o HEAD do branch default.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualScanResponse'
              example:
                status: queued
                commit_sha: 9b2e1f0a
                branch: main
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Repositorio nao encontrado.
        '409':
          description: Repositorio desativado, ou scan já em andamento para o commit.
          content:
            application/json:
              examples:
                desativado:
                  summary: Repositório desativado
                  value:
                    detail: 'Repositorio desativado: reative antes de iniciar um scan.'
                em_andamento:
                  summary: Scan já rodando para o mesmo commit
                  value:
                    detail: Ja existe um scan em andamento para o commit 9b2e1f0a.
        '502':
          description: Falha ao consultar o GitHub para resolver o HEAD do branch.
        '503':
          description: GitHub App nao configurado.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /repositories/{repository_id}/scans:
    get:
      tags:
      - repositories
      summary: Listar scans do repositório
      description: Lista os scans (ScanJob) do repositório. 404 se não pertencer ao usuário.
      operationId: list_repository_scans_repositories__repository_id__scans_get
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Repository Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanJobPage'
        '401':
          description: Token ausente, invalido ou expirado.
          content:
            application/json:
              example:
                detail: Credenciais invalidas ou token expirado.
        '404':
          description: Repositorio nao encontrado.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    FindingsSummary:
      properties:
        by_severity:
          additionalProperties:
            type: integer
          type: object
          title: By Severity
        by_tier:
          additionalProperties:
            type: integer
          type: object
          title: By Tier
        total:
          type: integer
          title: Total
      type: object
      required:
      - by_severity
      - by_tier
      - total
      title: FindingsSummary
      description: 'Agregado de findings de um scan: por severidade, por tier e total.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ManualScanResponse:
      properties:
        status:
          type: string
          title: Status
        commit_sha:
          type: string
          title: Commit Sha
        branch:
          type: string
          title: Branch
      type: object
      required:
      - status
      - commit_sha
      - branch
      title: ManualScanResponse
      description: 'Aceite de um scan manual (`POST /repositories/{id}/scan`).


        No espirito da resposta do webhook (`{"status": "queued", "commit_sha": ...}`),

        acrescentando o `branch` cujo HEAD foi resolvido.'
...
    ScanJobPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ScanJobSummary'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: ScanJobPage
      description: Pagina de resultados da listagem de scans recentes.
    ScanJobResponse:
      properties:
        commit_sha:
          type: string
          title: Commit Sha
        repo_url:
          type: string
          title: Repo Url
        repo_full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Repo Full Name
        pr_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pr Number
        tier1_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Tier1 Status
        tier1_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier1 Started At
        tier1_completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier1 Completed At
        tier2_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Tier2 Status
        tier2_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier2 Started At
        tier2_completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier2 Completed At
        tier3_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Tier3 Status
        tier3_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier3 Started At
        tier3_completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier3 Completed At
        blocked_at_tier:
          anyOf:
          - type: integer
          - type: 'null'
          title: Blocked At Tier
        final_risk_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Final Risk Score
        final_risk_level:
          anyOf:
          - type: string
          - type: 'null'
          title: Final Risk Level
        created_at:
          type: string
          format: date-time
          title: Created At
        findings_summary:
          $ref: '#/components/schemas/FindingsSummary'
      type: object
      required:
      - commit_sha
      - repo_url
      - repo_full_name
      - pr_number
      - tier1_status
      - tier1_started_at
      - tier1_completed_at
      - tier2_status
      - tier2_started_at
      - tier2_completed_at
      - tier3_status
      - tier3_started_at
      - tier3_completed_at
      - blocked_at_tier
      - final_risk_score
      - final_risk_level
      - created_at
      - findings_summary
      title: ScanJobResponse
      description: Representacao completa do status de um `ScanJob`, com resumo de findings.
    ScanJobSummary:
      properties:
        commit_sha:
          type: string
          title: Commit Sha
        repo_url:
          type: string
          title: Repo Url
        repo_full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Repo Full Name
        pr_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pr Number
        tier1_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Tier1 Status
        tier1_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier1 Started At
        tier1_completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier1 Completed At
        tier2_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Tier2 Status
        tier2_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier2 Started At
        tier2_completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier2 Completed At
        tier3_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Tier3 Status
        tier3_started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier3 Started At
        tier3_completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Tier3 Completed At
        blocked_at_tier:
          anyOf:
          - type: integer
          - type: 'null'
          title: Blocked At Tier
        final_risk_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Final Risk Score
        final_risk_level:
          anyOf:
          - type: string
          - type: 'null'
          title: Final Risk Level
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - commit_sha
      - repo_url
      - repo_full_name
      - pr_number
      - tier1_status
      - tier1_started_at
      - tier1_completed_at
      - tier2_status
      - tier2_started_at
      - tier2_completed_at
      - tier3_status
      - tier3_started_at
      - tier3_completed_at
      - blocked_at_tier
      - final_risk_score
      - final_risk_level
      - created_at
      title: ScanJobSummary
      description: Versao enxuta de `ScanJobResponse` (sem resumo de findings), usada em listagens.
    ScanReportResponse:
      properties:
        tier:
          type: integer
          title: Tier
        report_markdown:
          type: string
          title: Report Markdown
        analysis_json:
          additionalProperties: true
          type: object
          title: Analysis Json
        degraded:
          type: boolean
          title: Degraded
        comment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Comment Id
        posted:
          type: boolean
          title: Posted
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - tier
      - report_markdown
      - analysis_json
      - degraded
      - comment_id
      - posted
      - created_at
      title: ScanReportResponse
      description: Representacao de um `ScanReport` (relatorio markdown de um commit/tier).
    ScanReportsResponse:
      properties:
        commit_sha:
          type: string
          title: Commit Sha
        reports:
          items:
            $ref: '#/components/schemas/ScanReportResponse'
          type: array
          title: Reports
      type: object
      required:
      - commit_sha
      - reports
      title: ScanReportsResponse
      description: Lista de relatorios (um por tier) associados a um commit.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...